requirements.txt
google-cloud-bigquery==3.19.0
python-dotenv==1.0.1
requests==2.31.0
anthropic==0.20.0
asyncio==3.12.2
aiohttp==3.9.3
google-cloud-storage==2.16.0